refactor: #687 company-results を hooks/components 分割する
データ生成・取得 hooks・カード UI を分離し、
page 向け export パスは維持する。マッチング計算のテストを追加する。
